Refactored WosClassification class from one class to 4: WosClassification acts as an orchestrator; FileHandler handles file related actions; FacultyDepartmentManager handles things like updating counts and sets; CategoryProcessor handles getting the categories out of files and making them a key in the category_counts dict, and updating files set. Will likely move the updating counts all to a general Utility class when I refactor the Utility class in utilities.py
